Changed requirements, added conda requirements, parser configs.
I modified the requirements.txt heavily as many of the packages are
dependencies of each other and result in overspecification.
The conda_requirements.txt needs to be trimmed down but it's a useful
snapshot of what the environment looked like with all necessary
dependencies and no conflicts.
Modded the way that the parser configurations work.
Added a comment to parsers/telegram